bacon: Disable doze and app standby
These are poor workarounds to a deeper issue in Android. Currently, these
features interfere significantly with UX and apps' abilities to provide push
notifications (e.g. Gmail does not receive emails in a timely manner), so it's
better to have them disabled.

added a framework and setting commit(by sultanxda) to allow users to disable doze power save mode on the go.